Decompose
