+++ subduction_redb_storage/Cargo.toml
@@ -16,6 +16,7 @@ rust-version.workspace = true
 
 [dependencies]
 future_form = { workspace = true }
+futures = { workspace = true }
 redb = { workspace = true }
 sedimentree_core = { workspace = true }
 subduction_core = { workspace = true }
@@ -27,7 +28,6 @@ tracing = { workspace = true, default-features = true }
 [dev-dependencies]
 bolero = { workspace = true }
 criterion = { workspace = true }
-futures = { workspace = true }
 sedimentree_fs_storage = { workspace = true }
 subduction_core = { workspace = true, features = ["std", "test_utils"] }
 tempfile = { workspace = true }
